refactor: SideBarTemplate에서 Sidebar로 메뉴 데이터를 전달하도록 리팩터링
